Moved initInteractive code to hook
